refactor(utils): organize folders
