+++ packages/utils/src/beacon.mjs
